models/research/requirements.txt
absl-py==0.11.0
astunparse==1.6.3
blinker==1.4
cachetools==4.2.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.0
cycler==0.10.0
Cython==0.29.21
flatbuffers==1.12
gast==0.3.3
grpcio==1.32.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
jeepney==0.4.3
Keras-Preprocessing==1.1.2
keyring==21.3.0
kiwisolver==1.3.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lvis==0.5.3
macaroonbakery==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
netifaces==0.10.4
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.2.0
Pillow==8.0.1
protobuf
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycocotools==2.0.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
python-debian==0.1.37
python-xlib==0.27
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
scipy==1.5.4
SecretStorage==3.1.2
setuptools
simplejson==3.17.0
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
tf-models-official
termcolor==1.1.0
tf-slim==1.1.0
urllib3==1.25.9
wadllib==1.3.4
Werkzeug==1.0.1
wrapt==1.12.1

